fix doctest spike_train_synchrony

+++ elephant/spike_train_synchrony.py
@@ -153,7 +153,7 @@ def spike_contrast(spiketrains, t_start=None, t_stop=None,
     >>> spiketrain_2 = StationaryPoissonProcess(rate=20*pq.Hz,
     ...     t_stop=1000*pq.ms).generate_spiketrain()
     >>> round(spike_contrast([spiketrain_1, spiketrain_2]),3)
-    0.419
+    np.float64(0.419)
 
     """
     if not 0. < bin_shrink_factor < 1.:
